+++ NEWS.md
@@ -1,3 +1,7 @@
+## formatters 0.3.3.12
+ * `nlines` now accepts both `colwidths` and `max_width`, defaulting to `NULL` for both
+ * `pag_indices_inner`, `find_pag` and `valid_pag` now accept `have_col_fnotes` and `div_height` arguments
+
 ## formatters 0.3.3.11
  * fix bug which lead to Roche/rtables#414 (`nlines("", <non-null>)` threw an error